chat_ui/requirements.txt
# Chat UI Requirements
streamlit>=1.30.0
anthropic>=0.39.0
python-dotenv>=1.0.0
